refactor: 인기글 도메인 분리 (DTO, Service, Controller, Repository)
