+++ src/main/java/swyp12/team9/server/domain/recommendation/controller/RecommendationApi.java
@@ -70,23 +70,6 @@ public interface RecommendationApi {
             @Parameter(description = "가져올 추천 콘텐츠 수", example = "10") @RequestParam(defaultValue = "10") @Min(1) @Max(50) int size
     );
 
-    @Operation(summary = "인기글 조회", description = """
-                    공개 링크를 조회수 합계 기준으로 인기순 조회합니다.
-                    - 기준: 링크별 전역 공개 조회수 (publicViewCount)
-                    - 정렬: publicViewCount 내림차순, 동률이면 linkId 내림차순
-                    - 커서 형식: publicViewCount:linkId (예: 120:45)
-                    """)
-    @ApiSpec(status = HttpStatus.OK, errors = {
-            ErrorCode.VALIDATION_ERROR
-    })
-    @GetMapping("/popular")
-    ApiResponse<PaginationUtils.Cursor.PageResponse<RecommendationResponse>> getPopularLinks(
-            @CurrentUserId(required = false) Long userId,
-            @Parameter(description = "커서 (첫 요청 시 null, 형식: publicViewCount:linkId)", example = "120:45")
-            @RequestParam(required = false) String cursor,
-            @Parameter(description = "가져올 인기 콘텐츠 수", example = "10")
-            @RequestParam(defaultValue = "10") @Min(1) @Max(50) int size
-    );
 
     @Operation(summary = "외부/공개 링크 조회수 증가", description = """
                     추천 탭이나 인기글 탭 등에서 링크 카드를 클릭했을 때 호출합니다.
